+++ changelog.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
  - added flake8 testing to travis-ci 
  - updated code to make python3 compatible
  - updated code for pep9 styling
+ - use setArt() to set ListItem icons as the icon= constructor is deprecated
 
 ### Removed